requirements.txt
gradio
pytesseract
Pillow
openai
easyocr
